pyproject.toml
[tool.poetry]
name = "python-booking"
version = "0.1.0"
description = "A booking subsytem"
authors = ["Stanislas H.B. Sodonon <stanislas.sodonon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.68.0"}
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.7b0"
flake8 = "^3.9.2"
mypy = "^0.910"
types-python-dateutil = "^0.1.6"
isort = "^5.9.3"
pytest-cov = "^2.12.1"

[tool.black]
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.pytest_cache
  | dist
  | build
  | docs
)/
'''
include = '\.pyi?$'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
